pyproject.toml
[project]
name = "signal-box"
version = "0.3.0"
description = "Webhook relay service for alerts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
signal-box = "app.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
